externalized initial fetching sequence to useInit for organization
